feat:404ページ改良とヘッダーのuse client最小限化
