+++ src/app/layout.js
@@ -14,8 +14,8 @@ export default function RootLayout({ children }) {
           <link rel="apple-touch-icon" sizes="76x76" href="/favicon/apple-touch-icon.png"/>
           <link rel="icon" type="image/png" sizes="32x32" href="/favicon/favicon-32x32.png"/>
           <link rel="icon" type="image/png" sizes="16x16" href="/favicon/favicon-16x16.png"/>
-          <link rel="manifest" href="/site.webmanifest"/>
-          <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5"/>
+          <link rel="manifest" href="/favicon/site.webmanifest"/>
+          <link rel="mask-icon" href="/favicon/safari-pinned-tab.svg" color="#5bbad5"/>
           <meta name="msapplication-TileColor" content="#da532c"/>
           <meta name="theme-color" content="#ffffff"/>
           <html lang="ja">
